Correções das Correções

+++ Entra21.Gerenciador.Hospital.Veterinario/Models/Pet.cs
@@ -6,8 +6,8 @@ namespace Entra21.Gerenciador.Hospital.Vet.Models
         public int Id { get; set; }
         public string Nome { get; set; }
         public int Idade { get; set; }
-        public double Peso { get; set; }
-        public double Altura { get; set; }
+        public decimal Peso { get; set; }
+        public decimal Altura { get; set; }
         public char Genero { get; set; }
 
         public Responsavel Responsavel { get; set; }
